+++ internal/apitest/types.go
@@ -1,0 +1,11 @@
+package apitest
+
+type Server struct {
+	URL string `json:"url"`
+}
+
+type TestResult struct {
+	Endpoint string
+	Method   string
+	Status   string
+}
